Remove commented out code.

+++ include/render/OffscreenFramebuffers.inl
@@ -221,20 +221,10 @@ namespace vulpes {
         createAttachmentDescription(2, VK_IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL, VK_ATTACHMENT_LOAD_OP_DONT_CARE, VK_ATTACHMENT_STORE_OP_STORE);
     }
 
-	//template<>
-	//inline void OffscreenFramebuffers<picking_framebuffer_t>::createAttachmentDescriptions() {
-
-	//	createAttachmentDescription(0, VK_IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL, VK_ATTACHMENT_LOAD_OP_CLEAR, VK_ATTACHMENT_STORE_OP_DONT_CARE);
-	//	createAttachmentDescription(1, VK_IMAGE_LAYOUT_COLOR_ATTACHMENT_OPTIMAL, VK_ATTACHMENT_LOAD_OP_DONT_CARE, VK_ATTACHMENT_STORE_OP_STORE);
-	//	createAttachmentDescription(2, VK_IMAGE_LAYOUT_DEPTH_STENCIL_ATTACHMENT_OPTIMAL, VK_ATTACHMENT_LOAD_OP_CLEAR, VK_ATTACHMENT_STORE_OP_DONT_CARE);
-	//	// just need to copy to buffer before end of renderpass (so in same series of rendering commands, using pipeline barriers)
-
-	//}
-
 	template<typename offscreen_framebuffer_type>
 	inline void OffscreenFramebuffers<offscreen_framebuffer_type>::createAttachmentReference(const size_t & attachment_idx, const VkImageLayout & final_attachment_layout) {
 
-		attachmentReferences.push_back(VkAttachmentReference{ attachment_idx, final_attachment_layout });
+		attachmentReferences.push_back(VkAttachmentReference{ static_cast<uint32_t>(attachment_idx), final_attachment_layout });
 
 	}
